+++ src/search/ddg_search.rs
@@ -4,57 +4,38 @@ use super::utils;
 use indexmap::IndexMap;
 use std::sync::Arc;
 
-const CONTENT_SEP_FIRST: &str = "<div class=text>";
-const CONTENT_SEP_FINAL: &str = "<div class=article-bottom";
-const GEEKSFORGEEKS_SITE: &str = "www.geeksforgeeks.org";
-const GEEKSFORGEEKS_PAGE_URL: &str = "https://www.geeksforgeeks.org/";
-const GEEKSFORGEEKS_INVALID: [&str; 7] = [
-    "https://www.geeksforgeeks.org/tag/",
-    "https://www.geeksforgeeks.org/category/",
-    "https://www.geeksforgeeks.org/basic/",
-    "https://www.geeksforgeeks.org/easy/",
-    "https://www.geeksforgeeks.org/medium/",
-    "https://www.geeksforgeeks.org/hard/",
-    "https://www.geeksforgeeks.org/expert/",
-];
-
 /// These are the errors the functions associated with GeeksForGeeks will return.
 ///
-/// * `NotGfgPage` - The given url does not correspond to a GeeksForGeeks page.
 /// * `InvalidRequest` - Reqwest returned an error when processing the request. This can be
 /// due to rate limiting, bad internet etc.
 /// * `InvalidResponseBody` - The response content you got back is corrupted, usually bad
 /// internet.
-/// * `InvalidPageContent` - Usually this means the content returned by the website is
-/// corrupted because it did return 200 OK.
 /// * `ErrorCode` - The website returned an error code
 /// * `DdgError` - error with getting results from DuckDuckGO. (ddg::DdgError)
 #[derive(Debug)]
-pub enum GfgError {
-    NotGfgPage,
+pub enum DdgSearchError {
     InvalidRequest(reqwest::Error),
     InvalidReponseBody(reqwest::Error),
-    InvalidPageContent,
     ErrorCode(reqwest::StatusCode),
     DdgError(ddg::DdgError),
 }
 
-/// Scrape articles from GeeksForGeeks
-pub struct GeeksForGeeks {
+/// Scrape pages returned by ddg
+pub struct DdgSearch {
     client: Arc<reqwest::Client>,
     ddg: ddg::Ddg,
 }
 
-impl GeeksForGeeks {
-    /// Create a new GeeksForGeeks instance with a custom client that generates UA (user-agent in
+impl DdgSearch {
+    /// Create a new DdgSearch instance with a custom client that generates UA (user-agent in
     /// order to avoid getting rate limited by DuckDuckGO).
     ///
     /// # Examples
     ///
     /// ```
-    /// use falion::search::geeksforgeeks;
+    /// use falion::search::ddg_search;
     ///
-    /// let se = geeksforgeeks::GeeksForGeeks::new();
+    /// let ddg_search = ddg_search::DdgSearch::new();
     /// ```
     pub fn new() -> Self {
         Self {
@@ -63,14 +44,14 @@ impl GeeksForGeeks {
         }
     }
 
-    /// Create a new StackExchange instance with a provided client.
+    /// Create a new DdgSearch instance with a provided client.
     /// Note: DuckDuckGO will limit your requests if you don't provide a user-agent.
     ///
     /// ```
-    /// use falion::search::geeksforgeeks;
+    /// use falion::search::ddg_search;
     /// use std::sync::Arc;
     ///
-    /// let se = geeksforgeeks::GeeksForGeeks::with_client(Arc::new(reqwest::Client::new()));
+    /// let ddg_search = ddg_search::DdgSearch::with_client(Arc::new(reqwest::Client::new()));
     /// ```
     #[allow(dead_code)]
     pub fn with_client(client: Arc<reqwest::Client>) -> Self {
@@ -80,34 +61,32 @@ impl GeeksForGeeks {
         }
     }
 
-    /// Get the contents of a GeeksForGeeks page inside a String.
+    /// Get the contents of a page inside a String.
     ///
     /// # Arguments
     ///
-    /// * `page_url` - The GeeksForGeeks absolute url, specifically like this
-    /// https://www.geeksforgeeks.org/*, to the page.
+    /// * `page_url` - The absolute url to the page.
     ///
     /// # Examples
     ///
     /// ```
     /// use falion::search::ddg;
-    /// use falion::search::geeksforgeeks;
+    /// use falion::search::ddg_search;
     ///
-    /// # async fn run() -> Result<(), geeksforgeeks::GfgError> {
+    /// # async fn run() -> Result<(), ddg_search::DdgSearchError> {
     /// let ddg = ddg::Ddg::new();
-    /// let gfg = geeksforgeeks::GeeksForGeeks::new();
-    /// let link = &ddg.get_links("Rust basics", Some("www.geeksforgeeks.org"), None, Some(1)).await.unwrap()[0];
+    /// let ddg_search = ddg_search::DdgSearch::new();
+    /// let link = &ddg.get_links("Rust basics", None, None, Some(1)).await.unwrap()[0];
     ///
-    /// let page_content = gfg.get_page_content(&link).await.unwrap();
+    /// let page_content = ddg_search.get_page_content(&link).await.unwrap();
     /// # Ok(())
     /// # }
     /// ```
     ///
     /// # Errors
     ///
-    /// returns geeksforgeeks::GfgError
+    /// returns ddg_search::DdgSearchError;
     ///
-    /// * `NotGfgPage` - The given url does not correspond to a GeeksForGeeks page.
     /// * `InvalidRequest` - Reqwest returned an error when processing the request. This can be
     /// due to rate limiting, bad internet etc.
     /// * `InvalidResponseBody` - The response content you got back is corrupted, usually bad
@@ -115,71 +94,55 @@ impl GeeksForGeeks {
     /// * `InvalidPageContent` - Usually this means the content returned by the website is
     /// corrupted because it did return 200 OK.
     /// * `ErrorCode` - The website returned an error code
-    pub async fn get_page_content(&self, page_url: &str) -> Result<String, GfgError> {
+    pub async fn get_page_content(&self, page_url: &str) -> Result<String, DdgSearchError> {
         // set term width
         let term_width: usize = match crossterm::terminal::size() {
             Ok(size) => size.0.into(),
             Err(_) => 100,
         };
 
-        // check if page URL is valid
-        for invalid in GEEKSFORGEEKS_INVALID {
-            if page_url.contains(invalid) {
-                return Err(GfgError::NotGfgPage);
-            }
-        }
-
-        // get GeeksForGeeks page
+        // get page
         let response_body = match self.client.get(page_url).send().await {
             Ok(res) => {
                 if res.status() != reqwest::StatusCode::OK {
-                    return Err(GfgError::ErrorCode(res.status()));
+                    return Err(DdgSearchError::ErrorCode(res.status()));
                 }
 
                 match res.text().await {
                     Ok(body) => body,
-                    Err(err) => return Err(GfgError::InvalidReponseBody(err)),
+                    Err(err) => return Err(DdgSearchError::InvalidReponseBody(err)),
                 }
             }
-            Err(err) => return Err(GfgError::InvalidRequest(err)),
+            Err(err) => return Err(DdgSearchError::InvalidRequest(err)),
         };
 
-        // get the article part
-        let article = match response_body.split_once(CONTENT_SEP_FIRST) {
-            Some(res_split) => match res_split.1.split_once(CONTENT_SEP_FINAL) {
-                Some(art) => art.0,
-                None => return Err(GfgError::InvalidPageContent),
-            },
-            None => return Err(GfgError::InvalidPageContent),
-        };
-
-        // return article
-        Ok(utils::html_to_text(article, term_width))
+        // return page
+        Ok(utils::html_to_text(&response_body, term_width))
     }
 
-    /// Search for GeeksForGeeks results using duckduckgo and a provided query. This function will
+    /// Search for results using duckduckgo and a provided query. This function will
     /// go through ALL of those results and crate a future for each one which will start getting
     /// the content asynchronously for ALL of them. Each of this Futures is associated with the
     /// title of the page and returned inside an IndexMap for preserved order.
     ///
-    /// PLEASE READ: While setting a limit is optional, doing 100 requests to GeeksForGeeks at once
-    /// will probably get you rate limited.
+    /// PLEASE READ: While setting a limit is optional, doing multiple requests to possibly the
+    /// same site at once will probably get you rate limited.
     ///
     /// # Arguments
     ///
     /// * `query` - The query to search for.
-    /// * `limit` - Optional, but doing 100 requests to GeeksForGeeks at once will probably get you
-    /// rate limited. A recommended value is something like 10 for enough results and still good
-    /// results.
+    /// * `limit` - Optional, but doing multiple requests to possibly the same site at once will
+    /// probably get you rate limited. A recommended value is something like 10 for enough results
+    /// and still good results.
     ///
     /// # Examples
     ///
     /// ```
-    /// use falion::search::geeksforgeeks;
+    /// use falion::search::ddg_search;
     ///
-    /// # async fn run() -> Result<(), geeksforgeeks::GfgError> {
-    /// let gfg = geeksforgeeks::GeeksForGeeks::new();
-    /// let page_content = gfg
+    /// # async fn run() -> Result<(), ddg_search::DdgSearchError> {
+    /// let ddg_search = ddg_search::DdgSearch::new();
+    /// let page_content = ddg_search
     ///     .get_multiple_pages_content("Rust basics", Some(1))
     ///     .await
     ///     .unwrap();
@@ -193,7 +156,7 @@ impl GeeksForGeeks {
     ///
     /// # Errors
     ///
-    /// returns geeksforgeeks::GfgError;
+    /// returns ddg_search::DdgSearchError;
     ///
     /// * `DdgError` - error with getting results from DuckDuckGO. (ddg::DdgError)
     ///
@@ -203,15 +166,14 @@ impl GeeksForGeeks {
         &self,
         query: &str,
         limit: Option<usize>,
-    ) -> Result<IndexMap<String, tokio::task::JoinHandle<Result<String, GfgError>>>, GfgError> {
+    ) -> Result<
+        IndexMap<String, tokio::task::JoinHandle<Result<String, DdgSearchError>>>,
+        DdgSearchError,
+    > {
         // get the links from duckduckgo
-        let links = match self
-            .ddg
-            .get_links(query, Some(GEEKSFORGEEKS_SITE), Some(true), limit)
-            .await
-        {
+        let links = match self.ddg.get_links(query, None, Some(true), limit).await {
             Ok(res) => res,
-            Err(err) => return Err(GfgError::DdgError(err)),
+            Err(err) => return Err(DdgSearchError::DdgError(err)),
         };
 
         // create a new IndexMap
@@ -238,9 +200,9 @@ impl GeeksForGeeks {
     }
 }
 
-impl Default for GeeksForGeeks {
+impl Default for DdgSearch {
     fn default() -> Self {
-        GeeksForGeeks::new()
+        DdgSearch::new()
     }
 }
 
@@ -254,25 +216,23 @@ mod tests {
     use std::time::Duration;
 
     #[tokio::test]
-    async fn test_get_gfg_page() {
+    async fn test_get_ddg_page() {
         // random sleep time to prevent rate limiting when testing
         thread::sleep(Duration::from_secs(rand::thread_rng().gen_range(0..5)));
 
         let client = Arc::new(search::utils::client_with_special_settings());
         let ddg = search::ddg::Ddg::with_client(Arc::clone(&client));
-        let gfg = GeeksForGeeks::with_client(Arc::clone(&client));
+        let ddg_search = DdgSearch::with_client(Arc::clone(&client));
 
         let link = ddg
-            .get_links(
-                "Rust basics",
-                Some("www.geeksforgeeks.org"),
-                Some(false),
-                Some(1),
-            )
+            .get_links("Rust basics", None, Some(true), Some(1))
             .await
             .unwrap();
 
-        let page_content = gfg.get_page_content(link.first().unwrap()).await.unwrap();
+        let page_content = ddg_search
+            .get_page_content(link.first().unwrap())
+            .await
+            .unwrap();
 
         assert!(!page_content.is_empty())
     }
@@ -280,15 +240,15 @@ mod tests {
     // NOTE: Enable this test only when really needed in order to prevent rate limit with the other
     // tests
     // #[tokio::test]
-    // async fn test_get_multiple_gfg_pages_content() {
+    // async fn test_get_multiple_ddg_pages_content() {
     //     // random sleep time to prevent rate limiting when testing
     //     thread::sleep(Duration::from_secs(rand::thread_rng().gen_range(0..5)));
     //
     //     // actual function
     //     let client = Arc::new(utils::client_with_special_settings());
-    //     let gfg = GeeksForGeeks::with_client(Arc::clone(&client));
+    //     let ddg_search = DdgSearch::with_client(Arc::clone(&client));
     //
-    //     let page_content = gfg
+    //     let page_content = ddg_search
     //         .get_multiple_pages_content("Rust basics", Some(1))
     //         .await
     //         .unwrap();